reflactor: rearrange plan.go into separate files
